refactor(dashboard): extract Mastio dashboard helpers (F-B-201 PR-1/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-201.md

Sprint F-B-201 PR-1 of 10. base=main. Mirrors the F-B-202 sprint
pattern on the Mastio dashboard side.

mcp_proxy/dashboard/router.py is a 5106-LOC god-object with 16 feature
areas in a single file. This PR starts the modularization by lifting
the three pure helpers used across the surface into a sibling module
so upcoming per-feature sub-routers can import them without dragging
the whole router.

  mcp_proxy/dashboard/_helpers.py (new, 83 LOC):
    _ctx                          standard template context builder
    _enforce_safe_outbound_url    SSRF guard (Wave B G1 audit 2026-05-11)
    _login_client_ip              best-effort client IP for login lockout

  mcp_proxy/dashboard/router.py (5106 -> 5052 LOC, -54 net):
    - Three helper bodies removed; breadcrumb comments at the old
      positions.
    - One new ``from mcp_proxy.dashboard._helpers import ...`` block.

Pattern parity:
  - Court already shipped this split in F-B-202 PR-1 (#839,
    app/dashboard/_helpers.py).
  - Mastio _template_env.py already exists (predates this sprint),
    so no template factory work needed here.

Behavioural delta: zero. The three helpers retain identical bodies;
only their physical location changed. No route signatures touched.

Tests: pytest tests/test_proxy_dashboard*.py tests/test_h9_login_lockout.py
tests/test_fb9_proxy_logout_csrf.py tests/test_proxy_local_password_toggle.py
tests/test_enrollment_dashboard.py tests/test_h4_validate_config_sweep.py
-n auto --dist=loadfile -q
  103 passed.

Next PR-2 (F-B-201 modularization): extract the auth surface
(/proxy/login + /proxy/logout + /proxy/register + /proxy/setup,
~7 routes ~250 LOC) into mcp_proxy/dashboard/auth_routes.py mirroring
the Court PR-2 (#841) pattern.

Router LOC trajectory target:
  Pre-Sprint  : 66 routes, 5106 LOC
  Post-PR-1 (this): 66 routes, 5052 LOC
  Target post-PR-10: ~3 routes (overview + landing), ~120 LOC aggregator
